PR Feedback: Moved LoggingInputStream.
